renamed social folder to more appropriate name "interaction Hub" (fixed imports)
