+++ frontend/components/shared/PlayerBar.tsx
@@ -3,8 +3,8 @@ import Link from 'next/link'
 import { Box, Flex, IconButton, Image, Text, Slider, SliderTrack, SliderFilledTrack, SliderThumb, useBreakpointValue, HStack } from '@chakra-ui/react'
 import { FaPlay, FaPause, FaVolumeUp, FaVolumeMute, FaStepForward, FaStepBackward } from 'react-icons/fa'
 import { TbRewindBackward10, TbRewindForward10 } from 'react-icons/tb'
-import Likes from '../social/Likes'
-import Bookmarks from '../social/Bookmarks'
+import Likes from '../interactionHub/Likes'
+import Bookmarks from '../interactionHub/Bookmarks'
 import { convertTime } from '../../utilities/commonUtils'
 import { usePalette } from 'color-thief-react'
 import EndpointHelper from '../../helpers/EndpointHelper'
@@ -12,9 +12,9 @@ import { usePlayer } from '../../utilities/PlayerContext'
 import { SaveWatchHistoryRequest } from '../../types/Requests'
 import PodcastHelper from '../../helpers/PodcastHelper'
 import PlayerMenu from '../playerbar/Menu'
-import ChatBotButton from '../social/buttons/ChatBotButton'
-import CommentButton from '../social/buttons/CommentButton'
-import BookmarksButton from '../social/buttons/BookmarksButton'
+import ChatBotButton from '../interactionHub/buttons/ChatBotButton'
+import CommentButton from '../interactionHub/buttons/CommentButton'
+import BookmarksButton from '../interactionHub/buttons/BookmarksButton'
 
 const PlayerBar = () => {
   const { state, dispatch, audioRef } = usePlayer()
